+++ cmake/Findglfw3.cmake
@@ -5,13 +5,10 @@ include(FindPackageHandleStandardArgs)
 
 find_package(glfw3 CONFIG REQUIRED)
 
-# Package on >=18.04 sets a target, 16.04 uses default vars
+# Package on >=18.04 sets a target
 if (TARGET glfw)
   get_target_property(GLFW3_LIBRARIES glfw LOCATION)
   get_target_property(GLFW3_INCLUDE_DIRS glfw INTERFACE_INCLUDE_DIRECTORIES)
-else()
-  set(GLFW3_INCLUDE_DIRS ${GLFW3_INCLUDE_DIR})
-  set(GLFW3_LIBRARIES ${GLFW3_LIBRARY})
 endif()
 
 find_package_handle_standard_args(glfw3